Move EdgeRef and NodeRef to enums.py
