+++ tests/test_machine_learning.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from tests.utils import advance_to_play_turn, build_initial_placements
 from catanatron.state import player_deck_replenish
-from catanatron.models.enums import ORE, Action, ActionType, WHEAT
+from catanatron.models.enums import ORE, Action, ActionType, WHEAT, NodeRef
 from catanatron.models.board import Board, get_edges
 from catanatron.models.map import (
     BASE_MAP_TEMPLATE,
@@ -13,7 +13,6 @@ from catanatron.models.map import (
     NUM_EDGES,
     NUM_NODES,
     CatanMap,
-    NodeRef,
 )
 from catanatron.game import Game
 from catanatron.models.map import number_probability
